requirements.txt
requests~=2.31.0
jsonschema~=4.18.6
allure-pytest~=2.13.2
pytest~=7.4.0